requirements.txt
onnxruntime==1.24.4
numpy>=1.24
Pillow>=10.0
torch>=2.1.0
